TIP-1222: move the indexable product in bundle as it's for infrastructure
